Rename Request{Pending,Finished}Web3 -> Request{Pending,Finished}
