+++ js/src/views/Signer/components/RequestPending/index.js
@@ -14,4 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-export default from './RequestPendingWeb3';
+export default from './requestPending';
